add `ui/mute-button` tests
